# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Investigation & Policing Assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DashboardStatsCasesByStatusItem = {
  status: string;
  count: number;
};

export type DashboardStatsCasesByTypeItem = {
  type: string;
  count: number;
};

export interface DashboardStats {
  totalCases: number;
  activeCases: number;
  openComplaints: number;
  highPriorityAlerts: number;
  suspectsTracked: number;
  resolvedThisMonth: number;
  casesByStatus: DashboardStatsCasesByStatusItem[];
  casesByType: DashboardStatsCasesByTypeItem[];
}

export interface ActivityItem {
  id: number;
  type: string;
  title: string;
  description: string;
  /** @nullable */
  severity?: string | null;
  createdAt: string;
}

export interface Case {
  id: number;
  title: string;
  type: string;
  status: string;
  priority: string;
  description: string;
  officerAssigned: string;
  /** @nullable */
  suspectIds?: string | null;
  /** @nullable */
  evidence?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CaseInput {
  title: string;
  type: string;
  priority: string;
  description: string;
  officerAssigned: string;
  notes?: string;
}

export interface CaseUpdate {
  title?: string;
  status?: string;
  priority?: string;
  description?: string;
  officerAssigned?: string;
  notes?: string;
  evidence?: string;
}

export interface Complaint {
  id: number;
  complainantName: string;
  contactInfo: string;
  description: string;
  category: string;
  urgency: string;
  status: string;
  /** @nullable */
  keywords?: string | null;
  /** @nullable */
  aiSummary?: string | null;
  /** @nullable */
  linkedCaseId?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface ComplaintInput {
  complainantName: string;
  contactInfo: string;
  description: string;
}

export interface ComplaintUpdate {
  status?: string;
  urgency?: string;
  category?: string;
  linkedCaseId?: number;
  notes?: string;
}

export interface ComplaintAnalysis {
  complaintId: number;
  urgency: string;
  category: string;
  keywords: string[];
  summary: string;
  recommendedAction: string;
  sentiment?: string;
}

export interface Suspect {
  id: number;
  name: string;
  /** @nullable */
  alias: string | null;
  riskLevel: string;
  status: string;
  /** @nullable */
  nationalId?: string | null;
  /** @nullable */
  knownAssociates: string | null;
  /** @nullable */
  socialMediaProfiles?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  caseIds?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface SuspectInput {
  name: string;
  alias?: string;
  riskLevel: string;
  nationalId?: string;
  socialMediaProfiles?: string;
  notes?: string;
}

export interface SuspectUpdate {
  name?: string;
  alias?: string;
  riskLevel?: string;
  status?: string;
  knownAssociates?: string;
  socialMediaProfiles?: string;
  notes?: string;
}

export type SuspectNetworkNodesItem = {
  id: number;
  name: string;
  riskLevel: string;
  status?: string;
};

export type SuspectNetworkEdgesItem = {
  source: number;
  target: number;
  relationship: string;
};

export interface SuspectNetwork {
  nodes: SuspectNetworkNodesItem[];
  edges: SuspectNetworkEdgesItem[];
}

export interface Alert {
  id: number;
  title: string;
  description: string;
  severity: string;
  status: string;
  source: string;
  /** @nullable */
  linkedCaseId?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface AlertInput {
  title: string;
  description: string;
  severity: string;
  source: string;
  linkedCaseId?: number;
}

export interface AlertUpdate {
  status?: string;
  description?: string;
}

export interface OsintRequest {
  target: string;
  targetType: string;
  context?: string;
}

export interface OsintResult {
  id: number;
  target: string;
  targetType: string;
  riskScore: number;
  summary: string;
  flags: string[];
  suspiciousActivity?: string[];
  networkConnections?: string[];
  recommendation?: string;
  createdAt: string;
}

export interface CrimePattern {
  id: number;
  month: string;
  crimeType: string;
  count: number;
  changePercent: number;
}

export interface CrimeHotspot {
  id: number;
  area: string;
  district: string;
  lat: number;
  lng: number;
  incidentCount: number;
  riskLevel: string;
  primaryCrimeType: string;
}

export type CrimePredictionsPredictionsItem = {
  crimeType: string;
  predictedCount: number;
  confidence: number;
  trend: string;
};

export interface CrimePredictions {
  period: string;
  predictions: CrimePredictionsPredictionsItem[];
  highRiskAreas: string[];
  recommendations: string[];
}

export type ListCasesParams = {
  status?: string;
  priority?: string;
  search?: string;
};

export type ListComplaintsParams = {
  urgency?: string;
  status?: string;
  category?: string;
};

export type ListSuspectsParams = {
  search?: string;
  riskLevel?: string;
};

export type ListAlertsParams = {
  severity?: string;
  status?: string;
};