# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Investigation & Policing Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cases:
    get:
      operationId: listCases
      tags: [cases]
      summary: List cybercrime cases
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Case"
    post:
      operationId: createCase
      tags: [cases]
      summary: Create a new cybercrime case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseInput"
      responses:
        "201":
          description: Case created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
  /cases/{id}:
    get:
      operationId: getCase
      tags: [cases]
      summary: Get case details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Case details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
        "404":
          description: Case not found
    patch:
      operationId: updateCase
      tags: [cases]
      summary: Update case details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseUpdate"
      responses:
        "200":
          description: Updated case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
    delete:
      operationId: deleteCase
      tags: [cases]
      summary: Delete a case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /cases/{id}/suspects:
    get:
      operationId: getCaseSuspects
      tags: [cases]
      summary: Get suspects linked to a case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Suspects linked to case
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Suspect"
  /complaints:
    get:
      operationId: listComplaints
      tags: [complaints]
      summary: List citizen complaints
      parameters:
        - name: urgency
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of complaints
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Complaint"
    post:
      operationId: createComplaint
      tags: [complaints]
      summary: Submit a citizen complaint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplaintInput"
      responses:
        "201":
          description: Complaint submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
  /complaints/{id}:
    get:
      operationId: getComplaint
      tags: [complaints]
      summary: Get complaint details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Complaint details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
        "404":
          description: Not found
    patch:
      operationId: updateComplaint
      tags: [complaints]
      summary: Update complaint status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplaintUpdate"
      responses:
        "200":
          description: Updated complaint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
  /complaints/{id}/analyze:
    post:
      operationId: analyzeComplaint
      tags: [complaints]
      summary: AI-powered NLP analysis of a complaint
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplaintAnalysis"
...
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List threat alerts
      parameters:
        - name: severity
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
    post:
      operationId: createAlert
      tags: [alerts]
      summary: Create a threat alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertInput"
      responses:
        "201":
          description: Alert created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
  /alerts/{id}:
    patch:
      operationId: updateAlert
      tags: [alerts]
      summary: Update alert status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertUpdate"
      responses:
        "200":
          description: Updated alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
...
components:
  schemas:
...
    Case:
      type: object
      required: [id, title, type, status, priority, description, officerAssigned, createdAt, updatedAt]
      properties:
        id: { type: integer }
        title: { type: string }
        type: { type: string }
        status: { type: string }
        priority: { type: string }
        description: { type: string }
        officerAssigned: { type: string }
        suspectIds: { type: ["string", "null"] }
        evidence: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    CaseInput:
      type: object
      required: [title, type, priority, description, officerAssigned]
      properties:
        title: { type: string }
        type: { type: string }
        priority: { type: string }
        description: { type: string }
        officerAssigned: { type: string }
        notes: { type: string }
    CaseUpdate:
      type: object
      properties:
        title: { type: string }
        status: { type: string }
        priority: { type: string }
        description: { type: string }
        officerAssigned: { type: string }
        notes: { type: string }
        evidence: { type: string }
    Complaint:
      type: object
      required: [id, complainantName, contactInfo, description, category, urgency, status, createdAt, updatedAt]
      properties:
        id: { type: integer }
        complainantName: { type: string }
        contactInfo: { type: string }
        description: { type: string }
        category: { type: string }
        urgency: { type: string }
        status: { type: string }
        keywords: { type: ["string", "null"] }
        aiSummary: { type: ["string", "null"] }
        linkedCaseId: { type: ["integer", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    ComplaintInput:
      type: object
      required: [complainantName, contactInfo, description]
      properties:
        complainantName: { type: string }
        contactInfo: { type: string }
        description: { type: string }
    ComplaintUpdate:
      type: object
      properties:
        status: { type: string }
        urgency: { type: string }
        category: { type: string }
        linkedCaseId: { type: integer }
        notes: { type: string }
    ComplaintAnalysis:
      type: object
      required: [complaintId, urgency, category, keywords, summary, recommendedAction]
      properties:
        complaintId: { type: integer }
        urgency: { type: string }
        category: { type: string }
        keywords: { type: array, items: { type: string } }
        summary: { type: string }
        recommendedAction: { type: string }
        sentiment: { type: string }
    Suspect:
      type: object
      required: [id, name, alias, riskLevel, status, knownAssociates, createdAt, updatedAt]
      properties:
        id: { type: integer }
        name: { type: string }
        alias: { type: ["string", "null"] }
        riskLevel: { type: string }
        status: { type: string }
        nationalId: { type: ["string", "null"] }
        knownAssociates: { type: ["string", "null"] }
        socialMediaProfiles: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        caseIds: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
...
    Alert:
      type: object
      required: [id, title, description, severity, status, source, createdAt, updatedAt]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        severity: { type: string }
        status: { type: string }
        source: { type: string }
        linkedCaseId: { type: ["integer", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    AlertInput:
      type: object
      required: [title, description, severity, source]
      properties:
        title: { type: string }
        description: { type: string }
        severity: { type: string }
        source: { type: string }
        linkedCaseId: { type: integer }
    AlertUpdate:
      type: object
      properties:
        status: { type: string }
        description: { type: string }
...