# lib/api-zod/src/generated/types/case.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Investigation & Policing Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface Case {
  id: number;
  title: string;
  type: string;
  status: string;
  priority: string;
  description: string;
  officerAssigned: string;
  /** @nullable */
  suspectIds?: string | null;
  /** @nullable */
  evidence?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}