# lib/api-zod/src/generated/types/complaint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Investigation & Policing Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface Complaint {
  id: number;
  complainantName: string;
  contactInfo: string;
  description: string;
  category: string;
  urgency: string;
  status: string;
  /** @nullable */
  keywords?: string | null;
  /** @nullable */
  aiSummary?: string | null;
  /** @nullable */
  linkedCaseId?: number | null;
  createdAt: string;
  updatedAt: string;
}