# lib/api-zod/src/generated/types/complaintAnalysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Investigation & Policing Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface ComplaintAnalysis {
  complaintId: number;
  urgency: string;
  category: string;
  keywords: string[];
  summary: string;
  recommendedAction: string;
  sentiment?: string;
}