# lib/api-zod/src/generated/types/suspect.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Investigation & Policing Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface Suspect {
  id: number;
  name: string;
  /** @nullable */
  alias: string | null;
  riskLevel: string;
  status: string;
  /** @nullable */
  nationalId?: string | null;
  /** @nullable */
  knownAssociates: string | null;
  /** @nullable */
  socialMediaProfiles?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  caseIds?: string | null;
  createdAt: string;
  updatedAt: string;
}